requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
nomic==3.0.0
qdrant-client==1.7.0
requests==2.31.0
